requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
imapclient>=2.2.0
scikit-learn>=0.24.2
tensorflow>=2.6.0
nltk>=3.6.2
pandas>=1.3.0
numpy>=1.21.0
requests>=2.26.0
python-whois>=0.8.0
tldextract>=3.1.0
joblib>=1.0.1
beautifulsoup4>=4.9.3
email-validator>=1.1.3
python-multipart>=0.0.5